refactor(channels): derive channel-binding builder types from the Zod schema
The conversation channel-binding wire shape is defined once by a Zod schema whose responseBody drives openapi.yaml generation and, in turn, the web client's generated types. The Slack binding builder previously hand-wrote its own copy of that shape, which could silently drift from the contract.

Make the Zod schema the single source of truth on the server side too:
- Move the channel-binding schemas into a shared leaf module messaging/channel-binding-schema.ts; the route imports them for its responseBody, so openapi.yaml regenerates byte-for-byte identically (drift check passes).
- Derive ChannelBindingMetadata via z.infer (a Pick of the schema) and have the Slack builder and the dispatch use it, replacing the hand-written type and the earlier open Record<string, unknown>.

Structural change only — the wire shape is unchanged (openapi --check and the channelBinding integration tests both pass). LUM-2598.
